+++ src/terrain.cpp
@@ -1157,7 +1157,7 @@ void drawTerrain(void)
 			yPos = world_coord(y*sectorSize+sectorSize/2);
 			distance = pow(player.p.x+world_coord(visibleTiles.x/2) - xPos, 2) + pow(player.p.z+world_coord(visibleTiles.y/2) - yPos, 2);
 			//debug(LOG_TERRAIN, "culling: %i,%i player x: %i, player z: %i, xPos: %f, yPos: %f, distance = %f, max = %f",  x,y,player.p.x, player.p.z, xPos, yPos, distance, pow(world_coord(terrainDistance),2));
-			if (distance > pow(world_coord(terrainDistance),2))
+			if (distance > pow((double)world_coord(terrainDistance), 2))
 			{
 				sectors[x*ySectors + y].draw = false;
 			}
